third_party/WebKit/LayoutTests/editing/deleting/delete-character-002-expected.txt

a\n|
PASS sample.value is "a"
PASS sample.selectionStart is 1
PASS sample.selectionEnd is 1


a\n\n|
PASS sample.value is "a\n"
PASS sample.selectionStart is 2
PASS sample.selectionEnd is 2


a\n\n\n|
PASS sample.value is "a\n\n"
PASS sample.selectionStart is 3
PASS sample.selectionEnd is 3


a|\n
PASS sample.value is "\n"
PASS sample.selectionStart is 0
PASS sample.selectionEnd is 0


a\n|\n
PASS sample.value is "a\n"
PASS sample.selectionStart is 1
PASS sample.selectionEnd is 1


a\n\n|\n
PASS sample.value is "a\n\n"
PASS sample.selectionStart is 2
PASS sample.selectionEnd is 2


a\n|\n\n
PASS sample.value is "a\n\n"
PASS sample.selectionStart is 1
PASS sample.selectionEnd is 1


a\n\n|\n\n
PASS sample.value is "a\n\n\n"
PASS sample.selectionStart is 2
PASS sample.selectionEnd is 2


a\n|b
PASS sample.value is "ab"
PASS sample.selectionStart is 1
PASS sample.selectionEnd is 1


a\n\n|b
PASS sample.value is "a\nb"
PASS sample.selectionStart is 2
PASS sample.selectionEnd is 2


a\n\n\n|b
PASS sample.value is "a\n\nb"
PASS sample.selectionStart is 3
PASS sample.selectionEnd is 3


a\nb|
PASS sample.value is "a\n"
PASS sample.selectionStart is 2
PASS sample.selectionEnd is 2


a\n\nb|
PASS sample.value is "a\n\n"
PASS sample.selectionStart is 3
PASS sample.selectionEnd is 3


a\n\n\nb|
PASS sample.value is "a\n\n\n"
PASS sample.selectionStart is 4
PASS sample.selectionEnd is 4


PASS successfullyParsed is true

TEST COMPLETE
